ANDROID: sched: WALT: Refactor cumulative runnable average fixup
When a runnable task window is rolled over, it's CPU cumulative
runnable average needs to be updated accordingly. Currently this
is handled in a common function for all scheduling classes. However
when CFS_BANDWIDTH is enabled, fair scheduling class needs special
handling. As a preparation step, add a callback in sched_class
for updating the cumulative runnable average. The subsequent patches
make use of this callback and implement a different method for
fair scheduling class.

Bug: 139071966
Change-Id: Iffa88ab89254932177b1ea4eeaf12ec305c1c9c6
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: Arindam Bhattacharjee <abhattacharjee717@gmail.com>